+++ 7-period/java-lotto-7-main/src/test/java/lotto/LottoTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
